Moved the xyz folder into the newly created dflowfm folder and updated the references.
